Move toVariableReference method into PlannerUtils
